Create HTML for a Sign In With Google button

Creates the HTML element for displaying a Sign In With Google button. Most of the time you probably don't want to use this directly, but through the UI module googleSignInUI() instead. That allows the server to decode and verify the token received from the authentication process.

googleSignInButton(inputId, options = list())

Arguments

inputId
The input slot that authentication information will be available in. Note that while you can have several buttons in an app, the inputId of the first one found will determine the name of the Shiny input where authentication information will be passed.

options
A list of customization options to control the appearance of the button. data- will be automatically appended to the names. See GSI reference for possible values.

References

https://developers.google.com/identity/gsi/web/guides/display-button#html

https://developers.google.com/identity/gsi/web/reference/html-reference#element_with_class_g_id_signin