Create a verifier for verifying ID tokens

Create a verifier for verifying ID tokens

gsi_verifier(client_ids = character(), public_keys = NULL)

Arguments

client_ids
A character vector of client IDs your app recognizes.

public_keys
A list of public keys tried for signature verification, or a function that returns them. Keys can be objects, or paths to read with openssl::read_pubkey(). If left NULL, fetch Google's current public keys from their API when verifying.

See also

Examples

gsi_verifier()

#> <gsi_verifier>
#> List of 3
#>  $ client_ids     : chr(0) 
#>  $ get_public_keys:function ()  
#>  $ issuers        : chr [1:2] "accounts.google.com" "https://accounts.google.com"