Create HTML for a Sign In With Google button

Creates the HTML element for displaying a Sign In With Google button. Most of the time you probably don't want to use this directly, but through the googleSignInUI() module instead.

googleSignInButton(inputId, options = list())

Arguments

inputId
The input slot that authentication information will be available in. See Details.

options
A list of customization options to control the appearance of the button. The names will be automatically prepended with "data-". See Examples for usage and Google's reference for possible values.

Details

The given inputId receives the encoded Google ID JWT value. Most of the time you may want to use googleSignInUI() paired with the server module googleSignInServer() instead to decode and verify the token.

Note that while you can have several buttons in an app, whenever new authentication information is received it gets passed to all the googleSignInButton() inputs in the app.

References

https://developers.google.com/identity/gsi/web/guides/personalized-button

https://developers.google.com/identity/gsi/web/guides/display-button#html

https://developers.google.com/identity/gsi/web/reference/html-reference#element_with_class_g_id_signin

See also

Examples

# Customizing the appearance
googleSignInButton(
  inputId = "google_jwt",
  options = list(
    theme = "filled_blue",
    shape = "pill",
    text = "continue_with"
  )
)

#> <div id="google_jwt" class="g_id_signin" data-theme="filled_blue" data-shape="pill" data-text="continue_with"></div>